Revert "Random fix"

+++ OpenGL/src/controller/KeyboardController.hpp
@@ -28,5 +28,5 @@ private:
   bool dab;
 
   bool spaceIsPressed;
-  uint32_t lastSpacePress;
+  float lastSpacePress;
 };
