+++ OpenGL/src/entity/Member.cpp
@@ -34,6 +34,6 @@ void Member::draw() {
 }
 
 
-void Member::animate(uint32_t dt) {
+void Member::animate(float dt) {
   node.lookAt(anim->computeAnim(dt));
 }
